+++ packages/networks/bitcoin-esplora/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-chains/bitcoin-esplora
 
+## 2.0.0
+
+### Patch Changes
+
+- fix submit api route
+- Updated dependencies
+  - @rosen-chains/abstract-chain@7.0.0
+  - @rosen-chains/bitcoin@2.0.0
+
 ## 1.0.0
 
 ### Major Changes