+++ packages/networks/cardano-koios/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-chains/cardano-koios-network
 
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-chains/abstract-chain@7.0.0
+  - @rosen-chains/cardano@7.0.0
+
 ## 7.0.0
 
 ### Major Changes